requirements.txt
tensorflow-gpu>=2.7.0
keras>=2.7.0